Bug 1681750 - Fix random GamepadPlatformService::RemoveChannelParent crashes r=haik
The most likely cause for the crashes seems like IPDL intermittently calling GamepadEventChannelParent::ActorDestroy() multiple times on the same object.

This adds a wrapper to stop that behavior (if it's the root cause), and also to fire off release assertions at several other potential causes.

Hopefully we see this crash signature disappear.

Differential Revision: https://phabricator.services.mozilla.com/D99720

UltraBlame original commit: 2d192d43277bb6231b9145572d0059c5964f9d96

+++ dom/gamepad/ipc/GamepadEventChannelParent.cpp
@@ -56,7 +56,13 @@ GamepadEventChannelParent::GamepadEventChannelParent() {
 void GamepadEventChannelParent::ActorDestroy(ActorDestroyReason aWhy) {
   AssertIsOnBackgroundThread();
 
-  GamepadPlatformService::RemoveChannelParent(this);
+  
+  
+  
+  if (!mShutdown) {
+    GamepadPlatformService::RemoveChannelParent(this);
+    mShutdown = true;
+  }
 }
 
 mozilla::ipc::IPCResult GamepadEventChannelParent::RecvVibrateHaptic(
